+++ R/main.R
@@ -6,7 +6,8 @@ Options:
 --workers=N         Number of workers to spawn [default: 2]
 --port=PORT         Port to use [default: 8888]
 --results-dir=PATH  Directory to store model results in
---prerun-dir=PATH   Directory to find prerun results in"
+--prerun-dir=PATH   Directory to find prerun results in
+--inputs-dir=PATH   Directory to find input files in"
 
   validate_path <- function(path) {
     if (is.null(path)) {
@@ -20,13 +21,13 @@ Options:
        queue_id = dat$queue_id,
        workers = as.integer(dat$workers),
        results_dir = validate_path(dat[["results_dir"]]),
-       prerun_dir = validate_path(dat[["prerun_dir"]]))
+       inputs_dir = validate_path(dat[["inputs_dir"]]))
 }
 
 main_api <- function(args = commandArgs(TRUE)) {
   # nocov start
   dat <- main_api_args(args)
-  api <- api(dat$queue_id, dat$workers, dat$results_dir, dat$prerun_dir)
+  api <- api(dat$queue_id, dat$workers, dat$results_dir, dat$inputs_dir)
   api$run(host = "0.0.0.0", port = dat$port)
   # nocov end
 }
@@ -49,22 +50,17 @@ main_worker <- function(args = commandArgs(TRUE)) {
   if (args$calibrate_only) {
     worker_config <- "calibrate_only"
   }
-  worker <- rrq::rrq_worker_from_config(hintr_queue_id(args$queue_id, TRUE),
-                                        worker_config = worker_config)
+  worker <- rrq_worker_from_config(hintr_queue_id(args$queue_id, TRUE),
+                                    worker_config = worker_config)
   worker$loop()
   invisible(TRUE)
   # nocov end
 }
 
-main_import_prerun <- function(args = commandArgs(TRUE)) {
-  usage <- "Usage:
-  hintr-import-prerun <prerun> <path> [options]
-
-Options:
---model-output=PATH    Path to model output file [default: model-output.qs]"
-  args <- docopt_parse(usage, args)
-  h <- prerun_import(args$prerun, args$path, args$model_output)
-  message(sprintf("Imported data as '%s'", h))
+rrq_worker_from_config <- function(...) {
+  # nocov start
+  rrq::rrq_worker_from_config(...)
+  # nocov end
 }
 
 docopt_parse <- function(usage, args) {