+++ tests/testthat/test-03-run-model.R
@@ -1,5 +1,3 @@
-context("run-model")
-
 test_that("model can be run & calibrated and filters extracted", {
   test_mock_model_available()
   res <- process_result(mock_calibrate)
@@ -79,7 +77,8 @@ test_that("model without national level results can be processed", {
                c("area_id", "sex", "age_group", "calendar_quarter",
                  "indicator", "mode", "mean", "lower", "upper"))
   expect_true(nrow(res$data) > 84042)
-  expect_equivalent(as.data.frame(res$data)[1, "area_id"], "MWI_1_1_demo")
+  expect_equal(as.data.frame(res$data)[1, "area_id"], "MWI_1_1_demo",
+               ignore_attr = TRUE)
   expect_equal(names(res$plottingMetadata), c("barchart", "choropleth"))
   barchart <- res$plottingMetadata$barchart
   expect_equal(names(barchart), c("indicators", "filters", "defaults"))
@@ -246,4 +245,52 @@ test_that("real model can be run with csv2 data", {
   output <- naomi::read_hintr_output(model_run$model_output_path)
   expect_setequal(names(output),
                   c("output_package", "naomi_data", "info", "warnings"))
+})
+
+test_that("mock model can be forced to error", {
+  data <- list(
+    pjnz = file.path("testdata", "Malawi2019.PJNZ"),
+    shape = file.path("testdata", "malawi.geojson"),
+    population = file.path("testdata", "population.csv"),
+    survey = file.path("testdata", "survey.csv"),
+    programme = file.path("testdata", "programme.csv"),
+    anc = file.path("testdata", "anc.csv")
+  )
+  options <- list(
+    mock_model_trigger_error = "true",
+    area_scope = "MWI",
+    area_level = 4,
+    calendar_quarter_t1 = "CY2016Q1",
+    calendar_quarter_t2 = "CY2018Q3",
+    calendar_quarter_t3 = "CY2019Q2",
+    survey_prevalence = c("DEMO2016PHIA", "DEMO2015DHS"),
+    survey_art_coverage = "DEMO2016PHIA",
+    survey_recently_infected = "DEMO2016PHIA",
+    include_art_t1 = "true",
+    include_art_t2 = "true",
+    anc_clients_year2 = 2018,
+    anc_clients_year2_num_months = "9",
+    anc_prevalence_year1 = 2016,
+    anc_prevalence_year2 = 2018,
+    anc_art_coverage_year1 = 2016,
+    anc_art_coverage_year2 = 2018,
+    spectrum_population_calibration = "none",
+    spectrum_plhiv_calibration_level = "none",
+    spectrum_plhiv_calibration_strat = "sex_age_group",
+    spectrum_artnum_calibration_level = "none",
+    spectrum_artnum_calibration_strat = "age_coarse",
+    spectrum_infections_calibration_level = "none",
+    spectrum_infections_calibration_strat = "age_coarse",
+    calibrate_method = "logistic",
+    artattend_log_gamma_offset = -4L,
+    artattend = "false",
+    output_aware_plhiv = "true",
+    rng_seed = 17,
+    no_of_samples = 20,
+    max_iter = 250
+  )
+  expect_error(
+    run_model(data, options, tempdir()),
+    "Mock model has errored because option 'mock_model_trigger_error' is TRUE"
+  )
 })