+++ tests/testthat/test-08-endpoints-model.R
@@ -1,5 +1,3 @@
-context("endpoints-model")
-
 test_that("endpoint model run queues a model run", {
   test_redis_available()
   test_mock_model_available()
@@ -181,17 +179,14 @@ test_that("erroring model run returns useful messages", {
   expect_equal(status$id, response$id)
 
   # Get the result
-  mock_id <- mockery::mock(scalar("fake_key"), cycle = TRUE)
-  with_mock("ids::proquint" = mock_id, {
-    get_model_result <- model_result(queue)
-    error <- expect_error(get_model_result(response$id))
-  })
+  get_model_result <- model_result(queue)
+  error <- expect_error(get_model_result(response$id))
 
   expect_equal(error$status_code, 400)
   expect_equal(names(error$data[[1]]), c("error", "detail", "key", "trace"))
   expect_equal(error$data[[1]]$error, scalar("MODEL_RUN_FAILED"))
   expect_equal(error$data[[1]]$detail, scalar("test error"))
-  expect_equal(error$data[[1]]$key, scalar("fake_key"))
+  expect_match(error$data[[1]]$key, "\\w+-\\w+-\\w+")
 
   trace <- vapply(error$data[[1]]$trace, identity, character(1))
   expect_match(trace[[1]], "^# [[:xdigit:]]+$")
@@ -311,7 +306,7 @@ test_that("failed cancel sends reasonable message", {
   expect_equal(error$data[[1]]$error, scalar("FAILED_TO_CANCEL"))
   expect_match(error$data[[1]]$detail,
                scalar("Task [[:xdigit:]]+ is not cancelable \\(MISSING\\)"))
-  expect_is(error$data[[1]]$key, "character")
+  expect_type(error$data[[1]]$key, "character")
   expect_equal(error$status_code, 400)
 })