+++ tests/testthat/test-endpoints.R
@@ -1,12 +1,10 @@
-context("endpoints")
-
 test_that("plumber api can be built", {
   api <- api_build(NULL)
   expect_s3_class(api, "Plumber")
-  expect_length(api$routes, 10)
+  expect_length(api$routes, 11)
   expect_setequal(names(api$routes),
                  c("", "validate", "model", "calibrate", "comparison", "meta",
-                   "download", "hintr", "chart-data", "rehydrate"))
+                   "download", "hintr", "chart-data", "rehydrate", "internal"))
   expect_setequal(names(api$routes$validate),
                   c("baseline-individual", "baseline-combined",
                     "survey-and-programme", "options"))
@@ -18,13 +16,16 @@ test_that("plumber api can be built", {
   expect_equal(names(api$routes$`chart-data`), "input-time-series")
   expect_setequal(names(api$routes$rehydrate), c("submit", "status", "result"))
   expect_equal(names(api$routes$comparison), "plot")
+  expect_setequal(names(api$routes$internal), c("upload", "prerun"))
+  expect_setequal(names(api$routes$internal), c("upload", "prerun"))
+  expect_setequal(names(api$routes$internal$upload), c("input", "result"))
 })
 
 test_that("input_response correctly formats data and validates it", {
   mock_validate <- mockery::mock(TRUE)
   file <- list(path = "path", hash = "12345", filename = "original.pjnz", fromADR = FALSE,
                resource_url = "https://adr.unaids.org/file/123.pjnz")
-  with_mock("hintr:::validate_json_schema" = mock_validate, {
+  with_mock(validate_json_schema = mock_validate, {
     response <- input_response(list(data = list(country = scalar("Botswana")),
                                     filters = json_verbatim("null")),
                                "pjnz",
@@ -233,9 +234,9 @@ test_that("404 handler", {
   req <- list(REQUEST_METHOD = "POST",
               PATH_INFO = "/my/path")
   ans <- hintr_404_handler(req, res)
-  expect_is(ans, "list") # not json
+  expect_type(ans, "list") # not json
   expect_equal(ans$status, scalar("failure"))
-  expect_is(ans$errors[[1]]$key, "scalar")
+  expect_s3_class(ans$errors[[1]]$key, "scalar")
   expect_match(ans$errors[[1]]$key, "^[a-z]{5}-[a-z]{5}-[a-z]{5}$")
   # key is randomly generated - remove to compare rest
   ans$errors[[1]]$key <- NULL