+++ NEWS.md
@@ -1,7 +1,23 @@
-# hintr 1.1.3
+# hintr 1.1.10
 
 * Update fallback anc year to 2022
 * Remove stack trace from user display
+
+# hintr 1.1.9
+
+* Ensure Naomi options are translated based on `accept-language` header
+
+# hintr 1.1.8
+
+* Add `hintr_prerun` function which takes model inputs and outputs, saves them onto the server and then returns output zip with state JSON which can be used to rehydrate model outputs
+
+# hintr 1.1.7
+
+* Remove `prerun_import` and `prerun_push` functions, these are going to be replaced with a more transparent prerun method
+
+# hintr 1.1.6
+
+* Add endpoint `/internal/prerun` to run a prerun which builds state JSON from input and output files
 
 # hintr 1.1.2